[refactor]: put get_user_ai_models endpoint to ai_models router for consistency
